requirements.txt
comet-ml
easydict
numpy
python-dotenv
pytorch-lightning>=0.7.4
PyYAML
OpenEXR
Imath
scikit-image